Creation of Streamlit structure
