+++ streamlit/config.py
@@ -1,0 +1,30 @@
+"""
+
+Config file for Streamlit App
+
+"""
+
+from member import Member
+
+
+TITLE = "Le projet SatisPy"
+
+TEAM_MEMBERS = [
+    Member(
+        name="Quan",
+        linkedin_url="xxx",
+        github_url="xxx",
+    ),
+    Member(
+        name="Fred",
+        linkedin_url="xxx",
+        github_url="xxx",
+    ),
+
+    Member(
+        name="Eric GASNIER",
+        linkedin_url="https://www.linkedin.com/in/ericgasnier",
+    ),
+]
+
+PROMOTION = "Promotion ML Ops - Nov 2023"
